move some tests back
